requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.3
websockets==12.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiohttp==3.9.1
iqoptionapi==0.5
cryptography==41.0.7
requests==2.32.5
websocket-client==1.9.0
